requirements.txt
torch
PyYAML
torchvision
tqdm
lmdb
Pillow
tensorboardX
numpy